Moved User-related artifacts to user subpackage
