+++ src/main/java/com/abioye/rest/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.abioye.rest;
+package com.abioye.rest.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
